+++ draft-packages/form/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.2.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-form@7.2.8...@kaizen/draft-form@7.2.9) (2022-06-20)
+
+**Note:** Version bump only for package @kaizen/draft-form
+
+
+
+
+
 ## [7.2.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-form@7.2.7...@kaizen/draft-form@7.2.8) (2022-06-16)
 
 **Note:** Version bump only for package @kaizen/draft-form