+++ packages/loading-spinner/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/loading-spinner@2.2.1...@kaizen/loading-spinner@2.2.2) (2022-06-20)
+
+**Note:** Version bump only for package @kaizen/loading-spinner
+
+
+
+
+
 ## [2.2.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/loading-spinner@2.2.0...@kaizen/loading-spinner@2.2.1) (2022-05-30)
 
 **Note:** Version bump only for package @kaizen/loading-spinner